doctorUpdate and appointment mode update

+++ services/doctor-service.js
@@ -1,3 +1,4 @@
+import { Appointment } from "../models/appointment-model.js";
 import { Doctor } from "../models/doctor-model.js";
 import ApiError from "../utils/apiError.js";
 import { extractQuery } from "../utils/extractQuery.js";
@@ -40,6 +41,10 @@ export const updateDoctorService = async (id, data) => {
 
 export const deleteDoctorService = async (id) => {
 	if (!id) throw ApiError.notFound();
+	const checkDoctor = await Appointment.findOne({ doctor: id });
+	if (checkDoctor) {
+		throw ApiError.notAuthenticated("Sorry! Can not delete! This doctor is appointmens!")
+	}
 	const doctor = await Doctor.findByIdAndDelete(id);
 	return doctor;
 };
